utils/s3: split aws_error type registry into aws_error_definitions
Extract the `aws_error_type` enum and the `get_errors()` wire-name →
`aws_error` map out of aws_error.{hh,cc} into a new pair of files,
utils/s3/aws_error_definitions.{hh,cc}. aws_error.hh #includes the new
header; both build systems (configure.py and utils/CMakeLists.txt) get
the new .cc added to their source lists.

Preparation for auto-generating the enum and the map from the upstream
AWS SDK service models: keeping the registry in a dedicated translation
unit makes it possible to replace its S3/STS sections with generated
content without disturbing the rest of aws_error.

(cherry picked from commit c92e92239e849acca76116aec3402b74251b6550)
(cherry picked from commit 22d87d9f7632525c1ddf0070e1d8a3a10a24e314)
