+++ configure.py
@@ -1175,6 +1175,7 @@ scylla_core = (['message/messaging_service.cc',
                 'utils/http_client_error_processing.cc',
                 'utils/rest/client.cc',
                 'utils/s3/aws_error.cc',
+                'utils/s3/aws_error_definitions.cc',
                 'utils/s3/client.cc',
                 'utils/s3/default_aws_retry_strategy.cc',
                 'utils/s3/credentials_providers/aws_credentials_provider.cc',
